requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.10.0
pydantic-settings==2.7.0
redis==5.2.0
aiohttp==3.11.0
aiofiles==24.1.0
python-multipart==0.0.12
python-dotenv==1.0.1
lottie==0.7.2
